requirement.txt
earthengine-api 
geemap
pandas 
numpy 
scikit-learn 
pyyaml 
dvc
pandas
sklearn
mlflow